+++ src/main/java/org/codehaus/mojo/versions/UseNextReleasesMojo.java
@@ -76,14 +76,14 @@ public class UseNextReleasesMojo
         }
     }
 
-    private void useNextReleases( ModifiedPomXMLEventReader pom, Collection dependencies )
+    private void useNextReleases( ModifiedPomXMLEventReader pom, Collection<Dependency> dependencies )
         throws XMLStreamException, MojoExecutionException, ArtifactMetadataRetrievalException
     {
-        Iterator i = dependencies.iterator();
+        Iterator<Dependency> i = dependencies.iterator();
 
         while ( i.hasNext() )
         {
-            Dependency dep = (Dependency) i.next();
+            Dependency dep = i.next();
 
             if ( isExcludeReactor() && isProducedByReactor( dep ) )
             {